Populate cloud role and role instance tags
Cloud role is populated with the executable name of the current process
or "Rust Application" if getting the executable name failed.

Cloud role instance is populated using the machine hostname. This
replaces the previous implementation, which looked at the net.host.name
attribute.

Both of those changes align with the Azure Monitor exporters for Python
and JavaScript.