Fix #23.
The `duration_to_string` method has a small bug.

The padding for the microseconds was `7`; however, given the way the remaining micros are calculated, the remaining micros will only ever be a maximum of 6 digits.  In addition, microseconds in datetime strings should be formatted as exactly 6 digits.

Say you had `999999` microseconds remaining.  We would read this as `999 ms`.  However, using a padding of `7` makes this `0999999` microseconds.  I am guessing that app insights ignores the last digit, and consumes this as `099999` microseconds, or `99 ms`.

This should fix that issue.