Method to add service_name to all spans
It's useful to have a way to set the service name for all spans at once,
rather than having to set it every time.

This commit adds a with_service_name method to PipelineBuilder
and Exporter to allow users to set the service name in one place,
following the behaviour of the official DataDog and Zipkin exporters.

It also sets a default service_name if the user does not configure one.