gzip compression for requests
Application Insights supports gzip compression for requests to `POST
/v2/track`. When sending lots of telemetry items, network time starts
to be the limiting factor and compression helps with this.

Gzip compression is turned on by default and cannot be disabled. As far
as I can tell this is the same behaviour as the Application Insights
.NET SDK. If compression turns out to be an issue for some users, it
could be made optional in the future.